+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/misc/outertypefilename/package-info.java
