+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/outertypefilename/InputOuterTypeFilename1.java
@@ -1,5 +1,5 @@
 // someexamples of 1.5 extensions
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.misc.outertypefilename;
 
 public class InputOuterTypeFilename1 {
