+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/outertypefilename/InputOuterTypeFilenameIllegalTokens.java
@@ -1,0 +1,18 @@
+package com.puppycrawl.tools.checkstyle.checks.misc.outertypefilename;
+
+public class InputOuterTypeFilenameIllegalTokens
+{
+    public void methodWithPreviouslyIllegalTokens()
+    {
+        int i = 0;
+        switch (i)
+        {
+            default:
+                i--;
+                i++;
+                break;
+        }
+    }
+    
+    public native void nativeMethod();
+}
